feat(kernel): the concurrency token and the tenancy identifiers
Packet 6 step 2. The kernel and the schema disagreed in three places, and the
first migration cannot be written until they do not.

**The token is `long`, and one primitive advances it.** IOptimisticConcurrency
and AuditableEntity carried `uint` — the Npgsql convention for an xmin token,
which ADR-0039 rejected — against a `row_version bigint` column and a shipped
`EntityTag.For(long)` surface. Both are `long` now.

MarkUpdated and SoftDelete both route through a private Touch(at, by) that
stamps UpdatedAt/UpdatedBy and increments. SoftDelete used to assign the two
fields itself, so an increment placed in MarkUpdated alone would have left a
soft delete un-versioned and a client's pre-delete ETag would have kept
satisfying If-Match on the row it deleted. Mutation-checked both ways: reverting
SoftDelete fails SoftDelete_Advances_The_Row_Version and only that case;
dropping the increment fails two.

**The template's audit columns could not be satisfied.** `updated_by uuid NOT
NULL` against a nullable UpdatedBy that MarkCreated never stamps would have
rejected every INSERT; both are NULL now, with `coalesce(updated_at,
created_at)` named for last-touched. And `deleted_at` / `deleted_by` were listed
as a soft-delete opt-in while AuditableEntity implements ISoftDelete
unconditionally — EF maps them on every derived table, so a table that omitted
them could not materialize its own entity. They are unconditional; what is
opt-in is the query filter.

**TenantId and OrganizationId** in SharedKernel, per ADR-0023 Amendment 2's
cross-cutting placement: both appear on ITenantContext, on every marked entity,
in cache keys, job payloads and envelopes, so a module-owned type would make
each of those a reference to Tenancy. Neither has a New() — a tenant id is
assigned by the registry that owns the Tenant aggregate, because a handler that
minted its own could not satisfy the self-keyed policy's WITH CHECK.

I caught one of my own tests agreeing with the code rather than constraining it,
which is this packet's recurring lesson and worth recording. The first version
asserted the "canonical conversion mask" by round-tripping through STJ and the
TypeConverter. Measured: with the mask removed from TenantId entirely, all seven
cases still passed — Vogen's DEFAULT Conversions already emits both. What the
mask adds beyond the default is EfCoreValueConverter, so that is what the
assertion is on now, and the mutant dies.

The platform tenant sentinel is deliberately NOT added. The corpus asks for one
and fixes its value nowhere; the only consumer is audit_log, which Packet 9
owns. Choosing a one-way-door identifier for a table that does not exist is not
this step's call, and the absence is documented on the type rather than left to
be rediscovered.

737 tests green, 0 warnings under CI=true, format clean.

+++ backend/src/LearnStack.SharedKernel/Persistence/IOptimisticConcurrency.cs
@@ -27,5 +27,5 @@ public interface IOptimisticConcurrency
     /// update succeeds.
     /// </para>
     /// </remarks>
-    uint Version { get; }
+    long Version { get; }
 }
